Plugin addon-able BVH refactor.
